pyproject.toml
[tool.poetry]
name = "psk"
version = "0.1.0"
description = "Aplicación para mailing empresarial"
authors = ["SERGIO TEJEDOR MORENO <tejedor.moreno@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
dnspython = "2.3.0" # Esto lo añadi para solventar el problema con AttributeError: module 'httpcore' has no attribute 'NetworkBackend'
numpy = "^1.26.0"
streamlit = "^1.27.2"
yagmail = "^0.15.293"
langchain = "^0.0.310"
openai = "^0.28.1"
pandas = "^2.1.1"
keyring = "^24.2.0"
pymongo = "^4.5.0"
pydantic = "1.9.1"
streamlit-option-menu = "^0.3.6"
openpyxl = "^3.1.2"
python-dotenv = "^1.0.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
googletrans = "4.0.0-rc1"
icecream = "^2.1.3"
cryptography = "^41.0.4"
pytest = "^7.4.2"
pytest-cov = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
